add setting to disable "huge data" warning, closes #248
and be a bit more lenient for when the warning is shown (old limit was 1MB of data):
* either more than 5k "features"
* or ~10MB of data

the 5k map "features" limit might be hit relatively fast if one is using the "old" way to load OSM data (e.g. `way(x); out; >; out skel;`), as the way nodes also count as elements. but since `out geom` is the recommended way to fetch OSM data from overpass on overpass turbo since many years, this is an acceptable compromise.